Split up converter api in subconverts
